lowercase import name, moving types.
